auth fully implemented and routes protected

+++ backend/routes/plants.js
@@ -1,14 +1,15 @@
 const { Plant } = require("../models/plant")
-//const auth = require("../middleware/auth")
+const auth = require("../middleware/auth")
 const express = require("express")
 const Joi = require("joi")
 
 const router = express.Router()
 
-router.get('/', async (req, res) => {
+router.get('/', auth, async (req, res) => {
     try {
         const plants = await Plant.find()
-        res.send(plants)
+        const filteredPlants = plants.filter((plant) => plant.uid === req.user._id)
+        res.send(filteredPlants)
 
     } catch (error) {
         res.status(500).send(error.message)
@@ -16,7 +17,7 @@ router.get('/', async (req, res) => {
     }
 })
 
-router.post('/', async (req, res) => {
+router.post('/', auth, async (req, res) => {
     const schema = Joi.object({
         name: Joi.string().min(3).max(200).required(),
         author: Joi.string().min(3).max(30),
@@ -67,7 +68,7 @@ router.post('/', async (req, res) => {
     }
 })
 
-router.put('/:id', async (req, res) => {
+router.put('/:id', auth, async (req, res) => {
     const schema = Joi.object({
         name: Joi.string().min(3).max(200).required(),
         author: Joi.string().min(3).max(30),
@@ -87,13 +88,12 @@ router.put('/:id', async (req, res) => {
     const { error } = schema.validate(req.body)
     if (error) return res.status(400).send(error.details[0].message)
 
-
-
-
     try {
         const plant = await Plant.findById(req.params.id)
 
         if (!plant) return res.status(404).send("Plant does not exist in the database.")
+
+        if (plant.uid !== req.user._id) return res.status(401).send("Plant update failed. Not authorized.")
 
         const {
             name,
@@ -132,15 +132,16 @@ router.put('/:id', async (req, res) => {
     }
 })
 
-router.patch('/:id', async (req, res) => {
+router.patch('/:id', auth, async (req, res) => {
     try {
         const plant = await Plant.findById(req.params.id)
 
         if (!plant) return res.status(404).send("Plant does not exist in the database.")
 
+        if (plant.uid !== req.user._id) return res.status(401).send("Plant status change failed. Not authorized.")
+
         const updatedPlant = await Plant.findByIdAndUpdate(req.params.id, {
             isInGarden: !plant.isInGarden,
-
         })
         res.send(updatedPlant)
 
@@ -150,9 +151,16 @@ router.patch('/:id', async (req, res) => {
     }
 })
 
-router.delete('/:id', async (req, res) => {
+router.delete('/:id', auth, async (req, res) => {
     try {
+        const plant = await Plant.findById(req.params.id)
+
+        if (!plant) return res.status(404).send("Plant not found.")
+
+        if (plant.uid !== req.user._id) return res.status(401).send("Plant deletion failed. Not authorized.")
+
         const deletedPlant = await Plant.findByIdAndDelete(req.params.id)
+
         res.send(deletedPlant)
 
     } catch (error) {
